refactor: renames class
Renames helper class
